Add annotated monitor.asm
This contains an annotated asm file which includes documentation from
both the "Addressenboekje" as well as the "monitor.pdf"

Includes a series of notes. This is mainly used to figure out how to
get Mame to emulate the mini cassette.

Make sure the following always holds:

z80asm p2000.asm && diff a.bin p2000.rom && echo "All good"

To make sure the assembly file translates down to the original rom.