Reverse engineer block write function
Updates and annotates the block write function. The block write
function is now completely reverse engineered and explanations are added
on how the function behaves.

This should provide a good starting point for reverse engineering the
block read function, which follows a similar approach.

This was used to extract the checksum function, which is needed to
transform the existing .cas files into function .wav files that can be
consumed by mame.

Test: z80asm p2000.asm && diff a.bin p2000.rom && echo "all good"
     prints "all good"